requirements.txt
web3==6.10.0
pyyaml
requests
